\name{likelihood_lumpMLE}
\alias{likelihood_lumpMLE}
\title{Calculate likelihood of RDD and RT errors rates of STR using lumping MLE algorithm}
\usage{
likelihood_lumpMLE(Parameter_estimate,tempdataset,bin_size,N_core)
}
\description{
likelihood_lumpMLE will calculate the likelihood of RDD and RT errors rates of STR using lumping MLE algorithm
}
\examples{
likelihood_lumpMLE(Parameter_estimate,tempdataset,bin_size,N_core)
}
